% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_nwis_dv_dataset.R
\name{build_nwis_dv_dataset}
\alias{build_nwis_dv_dataset}
\title{Builds a date, streamflow, drainage area dataset of NWIS site data.}
\usage{
build_nwis_dv_dataset(sites, start_date, end_date)
}
\arguments{
\item{sites}{A character vector of NWIS sites.}

\item{start_date}{A string representation of the start date in YYYY-MM-DD format.}

\item{end_date}{A string representation of the end date in YYYY-MM-DD format.}
}
\description{
Takes a list of sites and a start and end date of interest. Returns
a named list containing dates, streamflow, and drainage area. Units are 
square miles and cubic feet per second.
}
\examples{
sites <- c("02177000","02178400")
start_date <- "2008-10-01"
end_date <- "2010-09-30"
build_nwis_dv_dataset(sites, start_date, end_date)
}
