% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_other_flow_stats.r
\name{calculate_other_flow_stats}
\alias{calculate_other_flow_stats}
\title{Function to return the OtherStats statistics for a given data series}
\usage{
calculate_other_flow_stats(flow_data, digits = 3)
}
\arguments{
\item{flow_data}{A dataframe containing a NWCCompare flow dataset. 
Should have been cleaned by \link[EflowStats]{validate_data}}

\item{digits}{A numeric. Number of digits to round indice values}
}
\value{
OtherStats data frame of calculated statistics
}
\description{
This is a function to compute the 7 statistics of daily streamflow. Input to the function is a
time series of streamflow with date in the format Y-m-d. Output is a vector of OtherStats 
1) mean, 2) median, 3) cv, 4) cv daily, 5) l7Q10, 6) l7Q2, 7) return_10 and 8) flow percentiles.
}
\examples{
library(EflowStats)
flow_data<-obs_data
flow_data$date <- as.Date(flow_data$date)
flow_data <- validate_data(flow_data, yearType = "water")
calculate_other_flow_stats(flow_data)
}
