% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_stat_cv}
\alias{calculate_stat_cv}
\title{Function to return the coefficient of variation for a given data series}
\usage{
calculate_stat_cv(x)
}
\arguments{
\item{x}{numeric vector of flow values}
}
\value{
numeric coefficient of variation for the given data frame
}
\description{
Function to return the coefficient of variation for a given data series
}
\details{
This functions accepts a numeric vector of flow values and calculates the coefficient of variation, defined as the mean/standard deviation
}
\examples{
calculate_stat_cv(obs_data$discharge)
}
