% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_stat_flow_perc}
\alias{calculate_stat_flow_perc}
\title{Function to return the requested flow percentiles for a given data series}
\usage{
calculate_stat_flow_perc(flow_data, probs = c(0.1, 0.25, 0.5, 0.75, 0.9,
  0.15))
}
\arguments{
\item{flow_data}{A dataframe containing a NWCCompare flow dataset. 
Should have been cleaned by \link[EflowStats]{validate_data}}

\item{probs}{vector containing requested percentile value(s)}
}
\value{
obs_percentiles requested flow percentiles for the given data frame
}
\description{
This function accepts a data frame containing daily data and a list of desired percentiles and 
returns a list of the requested percentiles
}
\examples{
library(EflowStats)
flow_data<-obs_data
flow_data$date <- as.Date(flow_data$date)
flow_data <- validate_data(flow_data, yearType = "water")
calculate_stat_flow_perc(flow_data, probs=c(.1,.25,.5,.75))

}
