% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_stat_nselog}
\alias{calculate_stat_nselog}
\title{Function to return the Nash-Sutcliffe value between the natural logarithms of the two data series}
\usage{
calculate_stat_nselog(estimate_timeseries, reference_timeseries, na.rm = TRUE)
}
\arguments{
\item{na.rm}{Boolean defaults to TRUE.}

\item{timeseries1}{data frame containing value data for one of the chosen timeseries}

\item{timeseries2}{data frame continaing value data for the second chosen timeseries}
}
\value{
nselog Nash-Sutcliffe value between the natural log of the two timeseries
}
\description{
This function accepts two data frames containing daily data series and returns the Nash-Sutcliffe value of the natural 
logarithms of the data, with zeros removed.
}
\examples{
obs_data<-obs_data
mod_data<-mod_data
calculate_stat_nselog(mod_data$discharge, obs_data$discharge)

}
