% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_stat_pbias}
\alias{calculate_stat_pbias}
\title{Function to return the percent bias between two data series}
\usage{
calculate_stat_pbias(timeseries2, timeseries1)
}
\arguments{
\item{timeseries2}{data frame continaing value data for the second chosen timeseries}

\item{timeseries1}{data frame containing value data for one of the chosen timeseries}
}
\value{
pbias percent bias between the two timeseries
}
\description{
This function accepts two data frames containing daily data series and returns the 
percent bias.
}
\examples{
timeseries1<-obs_data$discharge
timeseries2<-mod_data$discharge
calculate_stat_pbias(timeseries1,timeseries2)
}
