% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_stat_rmsne}
\alias{calculate_stat_rmsne}
\title{Function to return the normalized root mean square error between two data series}
\usage{
calculate_stat_rmsne(estimate_timeseries, reference_timeseries)
}
\arguments{
\item{estimate_timeseries}{data frame containing value data for the modeled timeseries}

\item{reference_timeseries}{data frame containing value data for the observed timeseries}
}
\value{
rmsne normalized root mean square error value between the two timeseries
}
\description{
This function accepts two data frames containing daily data series and returns 
the normalized root mean square error
}
\examples{
obs_data<-obs_data
mod_data<-mod_data
calculate_stat_rmsne(mod_data$discharge, obs_data$discharge)

}
