% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_stats_by_group.R
\name{calculate_stats_by_group}
\alias{calculate_stats_by_group}
\title{Function to calculate statistics of daily flow by grtoup for many sites}
\usage{
calculate_stats_by_group(stats, flow_data, yearType)
}
\arguments{
\item{stats}{string containing stat groups desired}

\item{flow_data}{A dataframe containing a NWCCompare flow dataset.}

\item{yearType}{A charcter of either "water" or "calendar" indicating 
whether to use water years or calendar years, respectively.}
}
\value{
Data frame of calculated statistics. Time series ids are used 
as row names, statistic ids as column names.
}
\description{
This function traverses a collection of daily streamflow data and 
calculate the requested statistics for each site. Available statistics
include those implemented by \link[EflowStats]{hitStats}, 
\link[EflowStats]{magnifSeven}, \link{functioname}
}
