% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_stats_diffs.R
\name{calculate_stats_diffs}
\alias{calculate_stats_diffs}
\title{Function to calculate the differences in statistics for two given datasets.}
\usage{
calculate_stats_diffs(sites, flow_data_a, flow_data_b, yearType = "water",
  digits = 3)
}
\arguments{
\item{sites}{A two column dataframe containing site names for flow_data_a
and flow_data_b flow data.}

\item{flow_data_a}{A dataframe containing a NWCCompare flow dataset. 
Should have been cleaned by \link[EflowStats]{validate_data} and constructed by
a build dataset function from this package.}

\item{flow_data_b}{A second NWCCompare flow dataset to be compared
to flow_data_a.}

\item{yearType}{A charcter of either "water" or "calendar" indicating 
whether to use water years or calendar years, respectively.}

\item{digits}{A numeric. Number of digits to round indice values}
}
\value{
statsout data frame of calculated statistics
}
\description{
This function accepts two data frames of daily flow data and returns a data frame of 
calculated difference statistics from \link[EflowStats]{calc_allHIT}, 
\link[EflowStats]{calc_magnifSeven}, \link{calculate_other_flow_stats}, and \link{calculate_GoF_stats}.
HitStats, calc_magnifSeven, and other_flow_stats are differented as \code{(a-b)/a}
}
\details{
Assumptions:
Dates must match between the two data sources.
}
\examples{
# https://cida.usgs.gov/nwc/#!waterbudget/achuc/031300011004
nwis <- "02335757"
huc <- "031300011004"
sites <- data.frame(a=nwis, b=huc, stringsAsFactors = FALSE)
start_date <- "2004-10-01"
end_date <- "2010-09-30"
flow_data_a <- build_nwis_dv_dataset(nwis, start_date, end_date)
flow_data_b <- build_nwc_flow_dataset(huc, start_date, end_date)
diff_statsout <- calculate_stats_diffs(sites = sites, 
  flow_data_a = flow_data_a, 
  flow_data_b = flow_data_b, 
  yearType = "water",
  digits = 2) 

}
