% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{find_peak_flow}
\alias{find_peak_flow}
\title{Function to find peak flow and date of peak flow in a time series.}
\usage{
find_peak_flow(x, choose = "no")
}
\arguments{
\item{x}{data frame with date, discharge, year_val, and day columns. 
Can be constructed with \link[EflowStats]{dataCheck}}

\item{choose}{character with options "first" and "last". If multiple 
equal peaks are found the function will choose the first or last if
this is set. Otherwise, an error is returned for multiple peaks.}
}
\value{
a \code{data.frame} with `peak` and `date` columns.
}
\description{
Accepts a \code{data.frame} as returned by \link[EflowStats]{dataCheck}
}
\examples{
library(EflowStats)
obs_data <- obs_data
obs_data$date <- as.Date(obs_data$date)
obs_data<-dataCheck(obs_data, yearType = "water")
find_peak_flow(obs_data)

}
