% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowStats_ICP.R
\name{FlowStatsICP}
\alias{FlowStatsICP}
\title{Function to run ICP HIT/HAT statistics for a given data set}
\usage{
FlowStatsICP(data, drain_area)
}
\arguments{
\item{data}{data frame of daily flow data}

\item{drain_area}{value of site drainage area}
}
\value{
Output data frame of calculated statistics
}
\description{
This function accepts a data frame of daily flow data and returns a data frame of 
calculated HIT/HAT statistics
}
\examples{
drain_area<-54
qfiletempf<-sampleData
qfiletempf$date<-as.Date(qfiletempf$date,"\%m/\%d/\%y")
FlowStatsICP(qfiletempf,drain_area)
}
