% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nwc_data.R
\name{get_nwc_huc}
\alias{get_nwc_huc}
\title{Get WFS polygon for a watershed.}
\usage{
get_nwc_huc(huc, local)
}
\arguments{
\item{huc}{The watershed of interest.}

\item{local}{TRUE/FALSE to request local watershed or total upstream watershed data.}
}
\value{
The data.
}
\description{
This function builds a request and returns a spatial polygons data frame.
}
\examples{
NWCwatershed<-get_nwc_huc(huc="031601030306",local=TRUE)

}
\author{
David Blodgett \email{dblodgett@usgs.gov}
}
