% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nwc_data.R
\name{get_nwc_wb_data}
\alias{get_nwc_wb_data}
\title{Get data from an NWC source.}
\usage{
get_nwc_wb_data(huc, local = FALSE)
}
\arguments{
\item{huc}{The watershed of interest.}

\item{local}{TRUE/FALSE to request local watershed or total upstream watershed data.}
}
\value{
The data.
}
\description{
This function builds a request and returns the the data in question.
}
\examples{
NWCdata<-get_nwc_wb_data(huc="031601030306")

}
\author{
David Blodgett \email{dblodgett@usgs.gov}
}
