% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nwc_data.R
\name{get_nwis_nwc_site}
\alias{get_nwis_nwc_site}
\title{Get an NWIS site that matches the HUC in question.}
\usage{
get_nwis_nwc_site(huc)
}
\arguments{
\item{huc}{The watershed of interest.}
}
\value{
The NWIS site ID. NULL if none found.
}
\description{
Given a HUC, this function checks if an NWIS site is available.
}
\examples{
data<-get_nwis_nwc_site(huc="031601030306")

}
\author{
David Blodgett \email{dblodgett@usgs.gov}
}
