% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obsdata.R
\name{get_obsdata}
\alias{get_obsdata}
\title{Function to transform data into the format required for calculation of HIT/HAT statistics}
\usage{
get_obsdata(x_obs)
}
\arguments{
\item{x_obs}{data frame pulled from dataRetrieval::readNWISdv or NWCCompare::get_nwc_wb_data}
}
\value{
obs_data data frame of observed data for the chosen parameters
}
\description{
This function accepts a raw data frame pulled from the NWIS webservice and returns a data frame of 
observed data for use in calculating HIT/HAT statistics
}
\examples{
site<-"02177000"
startdate<-"2011-10-01"
enddate<-"2012-09-30"
\dontrun{
x_obs <- dataRetrieval::readNWISdv(site)
obs_data <- get_obsdata(x_obs)
}
}
