\name{getSupportedStatNames}
\alias{getSupportedStatNames}
\title{Function to generate list of al possible names}
\usage{
getSupportedStatNames()
}
\value{
supportedNames list of all names supported by the package
}
\description{
This function accepts observed and modeled data frames of
daily flow data and returns a data frame of calculated diff
statistics
}

