% Generated by roxygen2 (4.0.1): do not edit by hand
\name{NWCportalReturn10}
\alias{NWCportalReturn10}
\title{Function to return the 10 year return value for a given data series}
\usage{
NWCportalReturn10(qfiletempf)
}
\arguments{
\item{qfiletempf}{data frame containing value data for the chosen timeseries}
}
\value{
NWCportalReturn10 10-year return value for the given data frame
}
\description{
This function accepts a data frame containing daily data and returns the 10 year return value,
calculated for the National Water Census by a non-standard method.
}
\examples{
qfiletempf<-sampleData
NWCportalReturn10(qfiletempf)
}

