\docType{data}
\name{obs_data}
\alias{obs_data}
\title{This is data to be included in my package}
\description{
  This is data to be included in my package
}
\author{
  Jessica Thompson \email{jlthomps@usgs.gov}

  Jessica Thompson \email{jlthomps@usgs.gov}

  Jessica Thompson \email{jlthomps@usgs.gov}
}
\keyword{data}
\keyword{datasets}

