% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherStatsICP.r
\name{OtherStatsICP}
\alias{OtherStatsICP}
\title{Function to return the OtherStats statistics for a given data series
This is a function to compute the 7 statistics of daily streamflow. Input to the function is a
time series of streamflow with date in the format Y-m-d. Output is a vector of OtherStats 
1) mean, 2) median, 3) cv, 4) cv daily, 5) l7Q10, 6) l7Q2, 7) return_10 and 8) flow percentiles.}
\usage{
OtherStatsICP(data)
}
\arguments{
\item{data}{data frame of daily flow data}
}
\value{
OtherStats data frame of calculated statistics
}
\description{
Function to return the OtherStats statistics for a given data series
This is a function to compute the 7 statistics of daily streamflow. Input to the function is a
time series of streamflow with date in the format Y-m-d. Output is a vector of OtherStats 
1) mean, 2) median, 3) cv, 4) cv daily, 5) l7Q10, 6) l7Q2, 7) return_10 and 8) flow percentiles.
}
\examples{
timeseries1<-sampleData
OtherStatsICP(timeseries1)
}
