% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_swe_csv.R
\name{parse_swe_csv}
\alias{parse_swe_csv}
\title{Function to return NWC Intercomparison portal modeled data for a given site}
\usage{
parse_swe_csv(input)
}
\arguments{
\item{input}{url for SOS service for desired site data}
}
\value{
data frame containing desired time series with names 'date' and 'data'
}
\description{
This function accepts a url and returns a data frame of data for that SOS request
}
\examples{
parse_swe_csv(system.file('extdata','SWECSVBlock_daymet_example.xml',package='NWCEd'))
}
