% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_nwc_flow_dataset.R
\name{build_nwc_flow_dataset}
\alias{build_nwc_flow_dataset}
\title{Builds a date, streamflow, drainage area dataset of NWC modeled data.}
\usage{
build_nwc_flow_dataset(hucs, start_date, end_date)
}
\arguments{
\item{hucs}{A character vector of 12-digit HUCs.}

\item{start_date}{A string representation of the start date in YYYY-MM-DD format.}

\item{end_date}{A string representation of the end date in YYYY-MM-DD format.}
}
\description{
Takes a list of hucs and a start and end date of interest. Returns
a named list containing dates, streamflow, and drainage area. Units are 
square miles and cubic feet per second.
}
\examples{
hucs <- c("031601020108","031501100104")
start_date <- "2008-10-01"
end_date <- "2010-09-30"
build_nwc_flow_dataset(hucs, start_date, end_date)
}
