% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_GoF_stats.R
\name{calculate_GoF_stats}
\alias{calculate_GoF_stats}
\title{Function to calculate Goodness of Fit statistics}
\usage{
calculate_GoF_stats(Modeled, Gaged)
}
\arguments{
\item{Modeled}{data frame of daily flow data}

\item{Gaged}{data frame of daily flow data}
}
\value{
Output data frame of calculated statistics
}
\description{
This function accepts observed and modeled data frames of daily flow 
data and returns a data frame of calculated GoF statistics
}
\examples{
Gaged <- obs_data
Gaged$date <- as.Date(Gaged$date)
Gaged <- dataCheck(Gaged, yearType = "water")
Modeled<-mod_data
Modeled$date <- as.Date(Modeled$date)
Modeled <- dataCheck(Modeled, yearType = "water")
calculate_GoF_stats(Modeled,Gaged)
}
