% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_stat_nse}
\alias{calculate_stat_nse}
\title{Function to return the Nash-Sutcliffe value between two data series}
\usage{
calculate_stat_nse(estimate_timeseries, reference_timeseries, na.rm = TRUE)
}
\arguments{
\item{estimate_timeseries}{data frame continaing value data for the second chosen timeseries}

\item{reference_timeseries}{data frame containing value data for one of the chosen timeseries}

\item{na.rm}{Boolean defaults to TRUE.}
}
\value{
nse Nash-Sutcliffe value between the two timeseries
}
\description{
This function accepts two data frames containing daily data series 
and returns the Nash-Sutcliffe value
}
\examples{
obs_data<-obs_data
mod_data<-mod_data
calculate_stat_nse(mod_data$discharge, obs_data$discharge)

}
