% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_stat_pbias}
\alias{calculate_stat_pbias}
\title{Function to return the percent bias between two data series}
\usage{
calculate_stat_pbias(estimate_timeseries, reference_timeseries)
}
\arguments{
\item{estimate_timeseries}{data frame containing value data for the modeled timeseries}

\item{reference_timeseries}{data frame containing value data for the observed timeseries}
}
\value{
pbias percent bias between the two timeseries
}
\description{
This function accepts two data frames containing daily data series and returns the 
percent bias.
}
\examples{
reference_timeseries<-obs_data$discharge
estimate_timeseries<-mod_data$discharge
calculate_stat_pbias(reference_timeseries,estimate_timeseries)
}
