% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_stat_rmsne}
\alias{calculate_stat_rmsne}
\title{Function to return the normalized root mean square error between two data series}
\usage{
calculate_stat_rmsne(timeseries1, timeseries2)
}
\arguments{
\item{timeseries1}{data frame containing value data for one of the chosen timeseries}

\item{timeseries2}{data frame continaing value data for the second chosen timeseries}
}
\value{
rmsne normalized root mean square error value between the two timeseries
}
\description{
This function accepts two data frames containing daily data series and returns the normalized root mean square error
}
\examples{
obs_data<-obs_data
mod_data<-mod_data
calculate_stat_rmsne(obs_data$discharge,mod_data$discharge)

}
