% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_stats_by_group.R
\name{calculate_stats_by_group}
\alias{calculate_stats_by_group}
\title{Function to calculate statistics of daily flow by group for many time series}
\usage{
calculate_stats_by_group(stats, flow_data, yearType = "water", digits = 3)
}
\arguments{
\item{stats}{string containing stat groups desired. 
options are: "all", "magAverage", "magLow", "magHigh", "frequencyLow", 
"frequencyHigh", "durationLow", "durationHigh", "timingAverage", 
"timingLow", "timingHigh", "rateChange", "magnifSeven", "otherStat"}

\item{flow_data}{A dataframe containing a NWCCompare flow dataset. 
Should have been cleaned by \link[EflowStats]{dataCheck}}

\item{yearType}{A charcter of either "water" or "calendar" indicating 
whether to use water years or calendar years, respectively.}

\item{digits}{A numeric. Number of digits to round indice values}
}
\value{
Data frame of calculated statistics. Time series ids are used 
as row names, statistic ids as column names.
}
\description{
This function traverses a collection of daily streamflow data and 
calculates the requested statistics for each time series. Available 
statistics include those implemented by \link[EflowStats]{hitAllStats}, 
\link[EflowStats]{magnifSeven}, \link{calculate_other_flow_stats}.
}
\examples{
sites <- c("02177000","02178400")
startdate <- "2008-10-01"
enddate <- "2013-09-30"
nwis_dataset <- build_nwis_dv_dataset(sites, startdate, enddate)
stats=c("hitMagAverage", "hitMagLow", "hitMagHigh", 
        "hitFrequencyLow", "hitFrequencyHigh", 
        "hitDurationLow", "hitDurationHigh", 
        "hitTimingAverage", "hitTimingLow", "hitTimingHigh", 
        "hitRateChange", 
        "magnifSeven", "otherStat")
eflow_stats <- calculate_stats_by_group(stats, nwis_dataset)
}
