% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_stats_diffs.R
\name{calculate_stats_diffs}
\alias{calculate_stats_diffs}
\title{Function to calculate the differences in  statistics for given observed and modeled data sets}
\usage{
calculate_stats_diffs(sites, startdate, enddate, X_DATA_FUN, x_args,
  DRAIN_AREA_FUN, drain_args, M_DATA_FUN, m_args)
}
\arguments{
\item{sites}{list of usgs station ids}

\item{startdate}{startdate for pulling data}

\item{enddate}{enddate for pulling data}

\item{X_DATA_FUN}{function for pulling data from x_args}

\item{x_args}{urls for pulling observed data}

\item{DRAIN_AREA_FUN}{function for pulling drainage area}

\item{drain_args}{url for pulling drainage area}

\item{M_DATA_FUN}{function for pulling modeled data form m_args}

\item{m_args}{url for pulling modeled data}
}
\value{
statsout data frame of calculated statistics
}
\description{
This function accepts observed and modeled data frames of daily flow data and returns a data frame of 
calculated diff statistics
}
