% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffInputs.R
\name{diffInputs}
\alias{diffInputs}
\title{Function to return the inputs for the calculate stats diff function}
\usage{
diffInputs(model_url)
}
\arguments{
\item{model_url}{url for access to modeled discharge data}
}
\value{
diffInputsv data frame of calculated statistics
}
\description{
Function to return the inputs for the calculate stats diff function
}
