\name{diffInputs}
\alias{diffInputs}
\title{Function to return the inputs for the calculate stats diff function}
\usage{
diffInputs(stats, model_url)
}
\arguments{
  \item{stats}{string of requested stats}

  \item{model_url}{url for access to modeled discharge
  data}
}
\value{
diffInputs data frame of calculated statistics
}
\description{
Function to return the inputs for the calculate stats diff
function
}

