% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nwc_data.R
\name{get_nwc_wb_data}
\alias{get_nwc_wb_data}
\title{Get data from an NWC source.}
\usage{
get_nwc_wb_data(huc, start_date = "1980-09-30", end_date = "2010-10-01",
  local = FALSE, return_var = "all")
}
\arguments{
\item{huc}{The watershed of interest.}

\item{start_date}{A string representation of the start date in YYYY-MM-DD format.}

\item{end_date}{A string representation of the end date in YYYY-MM-DD format.}

\item{local}{TRUE/FALSE to request local watershed or total upstream watershed data.}

\item{return_var}{A character vector specifying which variables are of interest.
Choose from "all", "et", "prcp", "discharge", and "streamflow". Discharge is
the modeled discharge for the HUC. Streamflow, if available is monitored streamflow 
near the outlet of the HUC.}
}
\value{
The data.
}
\description{
This function builds a request and returns the the data in question.
}
\examples{
NWCdata<-get_nwc_wb_data(huc="031601030306")

}
\author{
David Blodgett \email{dblodgett@usgs.gov}
}
