\name{getSupportedStatNames}
\alias{getSupportedStatNames}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Get Supported Stat Names}
\description{A convenience function to return currently supported statistics.}
\usage{getSupportedStatNames()}
\value{A list of statistic titles}
\author{David Blodgett dblodgett@usgs.gov}
\examples{getSupportedStatNames()}
