% Generated by roxygen2 (4.0.1): do not edit by hand
\name{l7Q2}
\alias{l7Q2}
\title{Function to return the l7Q2 value for a given data series}
\usage{
l7Q2(qfiletempf)
}
\arguments{
\item{qfiletempf}{data frame containing value data for one of the chosen timeseries}
}
\value{
l7Q2 l7Q2 value for the given data frame
}
\description{
This function accepts a data frame containing daily data and returns the l7Q2 value,
calculated for the National Water Census by a non-standard method.
}
\examples{
qfiletempf<-sampleData
l7Q2(qfiletempf)
}

