% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_other_flow_stats.r
\name{NWCportalL7Q2}
\alias{NWCportalL7Q2}
\title{Function to return the l7Q2 value for a given data series}
\usage{
NWCportalL7Q2(flow_data)
}
\arguments{
\item{flow_data}{data frame containing value data for one of the chosen timeseries}
}
\value{
NWCportalL7Q2 l7Q2 value for the given data frame
}
\description{
This function accepts a data frame containing daily data and returns the l7Q2 value, 
calculated for the National Water Census by a non-standard method.
}
\examples{
flow_data<-sampleData
NWCportalL7Q2(flow_data)
}
